LayoutTests/imported/w3c/web-platform-tests/svg/painting/parsing/marker-computed-expected.txt


FAIL Property marker value 'none' assert_equals: expected "none" but got ""
FAIL Property marker value 'url("https://example.com/")' assert_equals: expected "url(\"https://example.com/\")" but got ""
FAIL url values are made absolute assert_equals: expected "url(\"http://localhost:8800/svg/painting/parsing/a.b#c\")" but got ""
